remove unused classes and update tests
