+++ src/test/groovy/info/developia/gti/InjectorSpec.groovy
@@ -6,11 +6,9 @@ import spock.lang.Specification
 class InjectorSpec extends Specification {
     def "Injector should provide a instance of expected object"() {
         given:
-        Injector injector = new Injector(Library.class);
-
+        Injector injector = new Injector()
         when:
-        Library result = (Library) injector.getInstanceFor(Library.class);
-
+        Library result = (Library) injector.getInstanceOf(Library.class)
         then:
         result instanceof Library
     }
