+++ src/test/java/fixture/Launcher.java
@@ -1,12 +1,12 @@
 package fixture;
 
-import info.developia.gti.Gti;
+import info.developia.gti.Injector;
 
 public class Launcher {
+    private static final Injector injector = new Injector();
 
     public static void main(String[] args) {
-        Gti gti = new Gti();
-        Library library = (Library) gti.getInstanceOf(Library.class);
+        Library library = (Library) injector.getInstanceOf(Library.class);
         library.process();
     }
 }
